Added more types of AI
AI group one allows the mob to follow the player with "line of sight" from a line beem. I also added AI group two that follows a -x to x path and can JUMP! using the same method that allows the game to know if we're grounded or not. Instead of it checking if we're grouped it checks if the mob is near a jumpable block!