Fixed problems with AI and added funcationality to them (IMPORTANT NOTE ADDED READ)
So I made it now so the path of the walking/jumping mob will keep walking in one direction until he finds a wall, then he will switch directions and move  on the new path until he find another wall.

Something to keep in mind IMPORTANT!!!
The AI knows something is a wall if he runs into the base layer
The AI knows it has to jump if it runs into a jump layer. These are shown in the tile set layers if you want to modifiy the map!! The AI will NOT work if the layers are incorrect!